+++ common/processCap/ProcessCap.hpp
@@ -20,32 +20,15 @@
 #include <string>
 #include <sys/capability.h>
 
-namespace cta::server {
+namespace cta::server::ProcessCap {
 
-/**
- * Class providing support for UNIX capabilities.
- *
- * This class is used to provide support for UNIX capbilities, so that
- * subclasses can be created that override its virtual member functions.
- * Unit testing is the primary use-case where you may want a dummy capabilities
- * object that does nothing.
- *
- * Please note that process capabilities are not supported on Mac OS X.
- */
-class ProcessCap {
-public:
-  /**
-   * Destructor
-   */
-  virtual ~ProcessCap() = default;
-
-  /**
+   /**
    * C++ wrapper around the C functions cap_get_proc() and cap_to_text().
    *
    * @return The string representation the capabilities of the current
    * process.
    */
-  virtual std::string getProcText();
+  std::string getProcText();
 
   /**
    * C++ wrapper around the C functions cap_from_text() and cap_set_proc().
@@ -53,36 +36,5 @@ public:
    * @text The string representation the capabilities that the current
    * process should have.
    */
-  virtual void setProcText(const std::string &text);
-
-private:
-  /**
-   * C++ wrapper around the C function cap_get_proc().
-   *
-   * @return The capability state.
-   */
-  cap_t getProc();
-
-  /**
-   * C++ wrapper around the C function cap_to_text().
-   *
-   * @param cap The capability state.
-   */
-  std::string toText(const cap_t cap);
-
-  /**
-   * C++ wrapper around the C function cap_from_text().
-   *
-   * @return The capability state.
-   */
-  cap_t fromText(const std::string &text);
-
-  /**
-   * C++ wrapper around the C function cap_set_proc().
-   *
-   * @param cap The capability state.
-   */
-  void setProc(const cap_t cap);
-};
-
-} // namespace cta::server
+  void setProcText(const std::string &text);
+} // namespace cta::server::ProcessCap