+++ tapeserver/readtp/ReadtpCmd.cpp
@@ -187,16 +187,6 @@ std::list<std::string> ReadtpCmd::readListFromFile(const std::string &filename) 
 
 
 //------------------------------------------------------------------------------
-// setProcessCapabilities
-//------------------------------------------------------------------------------
-void ReadtpCmd::setProcessCapabilities(const std::string &capabilities) {
-  m_capUtils.setProcText(capabilities);
-  std::list<cta::log::Param> params;
-  params.push_back(cta::log::Param("capabilities", capabilities));
-  m_log(cta::log::DEBUG, "Set process capabilities", params);
-}
-
-//------------------------------------------------------------------------------
 // createDrive
 //------------------------------------------------------------------------------
 std::unique_ptr<castor::tape::tapeserver::drive::DriveInterface>