+++ ReleaseNotes.md
@@ -13,6 +13,7 @@
 - cta/CTA#645 - Fix new mount timeout log message
 - cta/CTA#666 - Fix drive status activity field not being properly reset
 - cta/CTA#502 - Fix string representation for Cleanup session type
+- cta/CTA#682 - Generate taped's log file with correct owner and group 
 
 ### Continuous Integration
 - cta/CTA#615 - Going to xrdfs xattr API for EOS5 extended attribute tests (EOS >= 5.2.17)